requirements.txt
# 在 requirements.txt 中添加以下内容
sentence-transformers==2.2.2
regex==2023.10.3
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography==41.0.8

# Milvus向量数据库
pymilvus==2.3.4

# 文档处理
PyMuPDF==1.23.18
python-docx==1.1.0
python-multipart==0.0.6
paddleocr==2.7.5
beautifulsoup4==4.12.2

# 机器学习与AI
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
text2vec==1.2.9
sentence-transformers==2.2.2
regex==2023.10.3

# 工具库
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
minio==7.2.0
pandas==2.1.4
numpy==1.25.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 监控和日志
prometheus-client==0.19.0
structlog==23.2.0